+++ utils/utils.py
@@ -1,0 +1,9 @@
+def custom_sort(data: dict, key: str) -> dict:
+    """
+    Custom sorts dict from key
+    """
+    return sorted(
+        data.items(),
+        reverse=True,
+        key=lambda item: item[-1][key]
+    )
